refactor(email-smtp): rename SMTP email injector to Email (SMTP) (#23)
- rename dir email/ -> email-smtp/ and package email_injector -> email_smtp_injector
- rename entrypoint openaev_email.py -> openaev_email_smtp.py
- change injector type openaev_email -> openaev_email_smtp and image to openaev/injector-email-smtp
- rename contract and injector display name to "Email (SMTP)"
- update manifest, docker-compose, .env/config samples and README

Distinguishes the SMTP transport from the new Email (Microsoft 365) and
Email (Google Workspace) API-based email injectors.
