+++ email-smtp/.dockerignore
@@ -1,0 +1,4 @@
+config.yml
+.env
+email_smtp_injector/__pycache__
+__pycache__
