+++ email-smtp/.env.sample
@@ -8,6 +8,6 @@ OPENAEV_TOKEN=ChangeMe
 OPENAEV_TENANT_ID=ChangeMe
 
 # INJECTOR Environment Variables
-INJECTOR_ID=email--ChangeMe
-INJECTOR_NAME="Email"
+INJECTOR_ID=email-smtp--ChangeMe
+INJECTOR_NAME="Email (SMTP)"
 INJECTOR_LOG_LEVEL=info
