+++ email-smtp/manifest-metadata.json
@@ -1,8 +1,8 @@
 {
-  "title": "Email",
-  "slug": "openaev_email",
-  "description": "Craft emails via SMTP",
-  "short_description": "Craft emails via SMTP",
+  "title": "Email (SMTP)",
+  "slug": "openaev_email_smtp",
+  "description": "Craft and send emails over SMTP",
+  "short_description": "Craft and send emails over SMTP",
   "use_cases": ["Technical"],
   "verified": true,
   "last_verified_date": "",
@@ -13,6 +13,6 @@
   "source_code": "",
   "manager_supported": true,
   "container_version": "rolling",
-  "container_image": "openaev/injector-email",
+  "container_image": "openaev/injector-email-smtp",
   "container_type": "INJECTOR"
 }
