--- email/.dockerignore
@@ -1,4 +1,0 @@
-config.yml
-.env
-email_injector/__pycache__
-__pycache__
